+++ demod_afsk.c
@@ -528,7 +528,7 @@ void demod_afsk_init (int samples_per_sec, int baud, int mark_freq,
           for (j=0; j<D->ms_filter_size; j++) {
 	    float am;
 	    float center;
-	    float shape = 1;		/* Shape is an attempt to smooth out the */
+	    float shape = 1.0f;		/* Shape is an attempt to smooth out the */
 					/* abrupt edges in hopes of reducing */
 					/* overshoot and ringing. */
 					/* My first thought was to use a cosine shape. */
@@ -536,16 +536,16 @@ void demod_afsk_init (int samples_per_sec, int baud, int mark_freq,
 					/* windows mentioned in the literature. */
 					/* http://en.wikipedia.org/wiki/Window_function */
 
-	    center = 0.5 * (D->ms_filter_size - 1);
-	    am = ((float)(j - center) / (float)samples_per_sec) * ((float)mark_freq) * (2 * M_PI);
+	    center = 0.5f * (D->ms_filter_size - 1);
+	    am = ((float)(j - center) / (float)samples_per_sec) * ((float)mark_freq) * (2.0f * (float)M_PI);
 
 	    shape = window (D->ms_window, D->ms_filter_size, j);
 
-	    D->m_sin_table[j] = sin(am) * shape;
-  	    D->m_cos_table[j] = cos(am) * shape;
+	    D->m_sin_table[j] = sinf(am) * shape;
+	    D->m_cos_table[j] = cosf(am) * shape;
 
-	    Gs += D->m_sin_table[j] * sin(am);
-	    Gc += D->m_cos_table[j] * cos(am);
+	    Gs += D->m_sin_table[j] * sinf(am);
+	    Gc += D->m_cos_table[j] * cosf(am);
 
 #if DEBUG1
 	    dw_printf ("%6d  %6.2f  %6.2f  %6.2f\n", j, shape, D->m_sin_table[j], D->m_cos_table[j]) ;
@@ -576,18 +576,18 @@ void demod_afsk_init (int samples_per_sec, int baud, int mark_freq,
           for (j=0; j<D->ms_filter_size; j++) {
 	    float as;
 	    float center;
-	    float shape = 1;
+	    float shape = 1.0f;
 
 	    center = 0.5 * (D->ms_filter_size - 1);
-	    as = ((float)(j - center) / (float)samples_per_sec) * ((float)space_freq) * (2 * M_PI);
+	    as = ((float)(j - center) / (float)samples_per_sec) * ((float)space_freq) * (2.0f * (float)M_PI);
 
 	    shape = window (D->ms_window, D->ms_filter_size, j);
 
-	    D->s_sin_table[j] = sin(as) * shape;
-  	    D->s_cos_table[j] = cos(as) * shape;
+	    D->s_sin_table[j] = sinf(as) * shape;
+	    D->s_cos_table[j] = cosf(as) * shape;
 
-	    Gs += D->s_sin_table[j] * sin(as);
-	    Gc += D->s_cos_table[j] * cos(as);
+	    Gs += D->s_sin_table[j] * sinf(as);
+	    Gc += D->s_cos_table[j] * cosf(as);
 
 #if DEBUG1
 	    dw_printf ("%6d  %6.2f  %6.2f  %6.2f\n", j, shape, D->s_sin_table[j], D->s_cos_table[j] ) ;
